# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "COVID-19 API Philippines DOH",
    "description": "COVID-19 API Philippines from DOH DATA DROP",
    "contact": {
      "name": "Github",
      "url": "https://github.com/Simperfy/Covid-19-API-Philippines-DOH"
    },
    "version": "0.9.0"
  },
  "paths": {
    "/api/filter/{field}/{value}": {
      "get": {
        "tags": ["Case Informations"],
        "deprecated": true,
        "operationId": "request_ziRnjRDBlc",
        "summary": "filters case informations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "WARNING",
                    "data",
                    "last_update",
                    "result_count"
                  ],
                  "properties": {
                    "WARNING": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "admitted",
                          "age",
                          "age_group",
                          "case_code",
                          "city_mun_res",
                          "city_muni_psgc",
                          "date_died",
                          "date_onset",
                          "date_recover",
                          "date_rep_conf",
                          "date_result_release",
                          "date_specimen",
                          "health_status",
                          "pregnant_tab",
                          "prov_res",
                          "quarantined",
                          "region_res",
                          "removal_type",
                          "sex"
                        ],
                        "properties": {
                          "date_died": {
                            "type": "string"
                          },
                          "city_mun_res": {
                            "type": "string"
                          },
                          "quarantined": {
                            "type": "string"
                          },
                          "date_onset": {
                            "type": "string"
                          },
                          "age_group": {
                            "type": "string"
                          },
                          "prov_res": {
                            "type": "string"
                          },
                          "city_muni_psgc": {
                            "type": "string"
                          },
                          "age": {
                            "type": "number"
                          },
                          "date_recover": {
                            "type": "string"
                          },
                          "sex": {
                            "type": "string"
                          },
                          "case_code": {
                            "type": "string"
                          },
                          "date_result_release": {
                            "type": "string"
                          },
                          "pregnant_tab": {
                            "type": "string"
                          },
                          "admitted": {
                            "type": "string"
                          },
                          "date_specimen": {
                            "type": "string"
                          },
                          "removal_type": {
                            "type": "string"
                          },
                          "region_res": {
                            "type": "string"
                          },
                          "date_rep_conf": {
                            "type": "string"
                          },
                          "health_status": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "last_update": {
                      "type": "string"
                    },
                    "result_count": {
                      "type": "number"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": [
        {
          "in": "path",
          "name": "field",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "in": "path",
          "name": "value",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/get": {
      "get": {
        "tags": ["Case Informations"],
        "operationId": "request_KcO1r9y92B",
        "summary": "Fetches case informations",
        "description": "Use to fetch all case informations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "last_update",
                    "pagination",
                    "result_count"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "admitted",
                          "age",
                          "age_group",
                          "case_code",
                          "city_mun_res",
                          "city_muni_psgc",
                          "date_died",
                          "date_onset",
                          "date_recover",
                          "date_rep_conf",
                          "date_result_release",
                          "date_specimen",
                          "health_status",
                          "pregnant_tab",
                          "prov_res",
                          "quarantined",
                          "region_res",
                          "removal_type",
                          "sex"
                        ],
                        "properties": {
                          "date_died": {
                            "type": "string"
                          },
                          "city_mun_res": {
                            "type": "string"
                          },
                          "quarantined": {
                            "type": "string"
                          },
                          "date_onset": {
                            "type": "string"
                          },
                          "age_group": {
                            "type": "string"
                          },
                          "prov_res": {
                            "type": "string"
                          },
                          "city_muni_psgc": {
                            "type": "string"
                          },
                          "age": {
                            "type": "number"
                          },
                          "date_recover": {
                            "type": "string"
                          },
                          "sex": {
                            "type": "string"
                          },
                          "case_code": {
                            "type": "string"
                          },
                          "date_result_release": {
                            "type": "string"
                          },
                          "pregnant_tab": {
                            "type": "string"
                          },
                          "admitted": {
                            "type": "string"
                          },
                          "date_specimen": {
                            "type": "string"
                          },
                          "removal_type": {
                            "type": "string"
                          },
                          "region_res": {
                            "type": "string"
                          },
                          "date_rep_conf": {
                            "type": "string"
                          },
                          "health_status": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "last_update": {
                      "type": "string"
                    },
                    "pagination": {
                      "type": "object",
                      "required": [
                        "limit",
                        "max_page",
                        "next_page"
                      ],
                      "properties": {
                        "limit": {
                          "type": "number"
                        },
                        "max_page": {
                          "type": "number"
                        },
                        "next_page": {
                          "type": "number"
                        }
                      }
                    },
                    "result_count": {
                      "type": "number"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": [
        {
          "name": "page",
          "description": "page for entries",
          "in": "query",
          "required": false,
          "type": "integer",
          "example": 1,
          "$ref": ""
        },
        {
          "name": "limit",
          "description": "limit for entries",
          "in": "query",
          "required": false,
          "type": "integer",
          "example": 10,
          "$ref": ""
        },
        {
          "name": "month",
          "description": "returns record with this month",
          "in": "query",
          "required": false,
          "type": "string",
          "example": "03",
          "$ref": ""
        },
        {
          "name": "day",
          "description": "returns record with this day",
          "in": "query",
          "required": false,
          "type": "string",
          "example": "30",
          "$ref": ""
        }
      ]
    },
    "/api/timeline": {
      "get": {
        "tags": ["Case Informations"],
        "operationId": "request_FSpSE0nMqK",
        "summary": "fetches covid-19 cases timeline",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "last_update"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "cases",
                          "date"
                        ],
                        "properties": {
                          "cases": {
                            "type": "number"
                          },
                          "date": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "last_update": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
    "/api/top-regions": {
      "get": {
        "tags": ["Case Informations"],
        "operationId": "request_STjLIaQLap",
        "summary": "fetches top regions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "last_update"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "cases",
                          "region"
                        ],
                        "properties": {
                          "cases": {
                            "type": "number"
                          },
                          "region": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "last_update": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
    "/api/summary": {
      "get": {
        "tags": ["Case Informations"],
        "operationId": "request_VPHfKLz3Mm",
        "summary": "fetch summary",
        "description": "fetches summary overall or by region",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "last_update"
                  ],
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string"
                    },
                    "last_update": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
    "/api/facilities": {
      "get": {
        "tags": ["Facility Informations"],
        "operationId": "request_sbDz8vhQHS",
        "summary": "fetches facilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "last_update"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "added_date",
                          "beds_ward_o",
                          "beds_ward_v",
                          "cf_name",
                          "city_mun",
                          "city_mun_psgc",
                          "conf_asym",
                          "conf_crit",
                          "conf_died",
                          "conf_mild",
                          "conf_severe",
                          "discharged",
                          "doctor_adm",
                          "east_coord",
                          "hfhudcode",
                          "icu_o",
                          "icu_o_nc",
                          "icu_v",
                          "icu_v_nc",
                          "isolbed_o",
                          "isolbed_v",
                          "mechvent_o",
                          "mechvent_o_nc",
                          "mechvent_v",
                          "mechvent_v_nc",
                          "nonicu_o_nc",
                          "nonicu_v_nc",
                          "north_coord",
                          "nurse_adm",
                          "other_adm",
                          "prob_asym",
                          "prob_crit",
                          "prob_died",
                          "prob_mild",
                          "prob_severe",
                          "province",
                          "province_psgc",
                          "q_doctor",
                          "q_nurse",
                          "q_other",
                          "region",
                          "region_psgc",
                          "report_date",
                          "susp_asym",
                          "susp_crit",
                          "susp_died",
                          "susp_mild",
                          "susp_severe",
                          "t_patient_adm",
                          "t_patient_er",
                          "t_patient_icu",
                          "trans_ttmf",
                          "updated_date"
                        ],
                        "properties": {
                          "conf_severe": {
                            "type": "number"
                          },
                          "susp_crit": {
                            "type": "number"
                          },
                          "icu_o": {
                            "type": "number"
                          },
                          "mechvent_o_nc": {
                            "type": "number"
                          },
                          "mechvent_o": {
                            "type": "number"
                          },
                          "other_adm": {
                            "type": "number"
                          },
                          "east_coord": {
                            "type": "string"
                          },
                          "mechvent_v_nc": {
                            "type": "number"
                          },
                          "report_date": {
                            "type": "string"
                          },
                          "city_mun_psgc": {
                            "type": "string"
                          },
                          "prob_severe": {
                            "type": "number"
                          },
                          "t_patient_icu": {
                            "type": "number"
                          },
                          "added_date": {
                            "type": "string"
                          },
                          "icu_o_nc": {
                            "type": "number"
                          },
                          "region_psgc": {
                            "type": "string"
                          },
                          "conf_mild": {
                            "type": "number"
                          },
                          "city_mun": {
                            "type": "string"
                          },
                          "prob_asym": {
                            "type": "number"
                          },
                          "susp_mild": {
                            "type": "number"
                          },
                          "susp_asym": {
                            "type": "number"
                          },
                          "t_patient_er": {
                            "type": "number"
                          },
                          "isolbed_v": {
                            "type": "number"
                          },
                          "north_coord": {
                            "type": "string"
                          },
                          "icu_v": {
                            "type": "number"
                          },
                          "updated_date": {
                            "type": "string"
                          },
                          "trans_ttmf": {
                            "type": "number"
                          },
                          "t_patient_adm": {
                            "type": "number"
                          },
                          "q_doctor": {
                            "type": "number"
                          },
                          "nonicu_v_nc": {
                            "type": "number"
                          },
                          "prob_mild": {
                            "type": "number"
                          },
                          "discharged": {
                            "type": "number"
                          },
                          "icu_v_nc": {
                            "type": "number"
                          },
                          "conf_asym": {
                            "type": "number"
                          },
                          "mechvent_v": {
                            "type": "number"
                          },
                          "susp_died": {
                            "type": "number"
                          },
                          "cf_name": {
                            "type": "string"
                          },
                          "beds_ward_o": {
                            "type": "number"
                          },
                          "q_other": {
                            "type": "number"
                          },
                          "province": {
                            "type": "string"
                          },
                          "region": {
                            "type": "string"
                          },
                          "conf_died": {
                            "type": "number"
                          },
                          "isolbed_o": {
                            "type": "number"
                          },
                          "doctor_adm": {
                            "type": "number"
                          },
                          "prob_crit": {
                            "type": "number"
                          },
                          "prob_died": {
                            "type": "number"
                          },
                          "province_psgc": {
                            "type": "string"
                          },
                          "susp_severe": {
                            "type": "number"
                          },
                          "hfhudcode": {
                            "type": "string"
                          },
                          "conf_crit": {
                            "type": "number"
                          },
                          "q_nurse": {
                            "type": "number"
                          },
                          "beds_ward_v": {
                            "type": "number"
                          },
                          "nonicu_o_nc": {
                            "type": "number"
                          },
                          "nurse_adm": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "last_update": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    }
  },
  "components": {
    "schemas": {}
  }
}